Moved hssp to a separate file
