+++ optuna/_hypervolume/__init__.py
@@ -1,7 +1,7 @@
 from optuna._hypervolume.base import BaseHypervolume
+from optuna._hypervolume.hssp import _solve_hssp
 from optuna._hypervolume.utils import _compute_2d
 from optuna._hypervolume.utils import _compute_2points_volume
-from optuna._hypervolume.utils import _solve_hssp
 from optuna._hypervolume.wfg import WFG
